# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM API Documentation",
        "description": "REST API for managing leads, customers, notes, tasks, activities, and dashboard statistics. Authentication uses JWT bearer tokens obtained via the login or register endpoints.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List tasks",
                "description": "Returns a paginated, filterable list of tasks. Accessible to admins, sales managers, and sales executives.",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by assigned user ID.",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by task status.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter by task priority.",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high"
                            ]
                        }
                    },
                    {
                        "name": "overdue",
                        "in": "query",
                        "description": "Filter to only overdue tasks.",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page.",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of tasks.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden for the current role.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a task",
                "description": "Creates a new task. A task must belong to either a lead or a customer. Accessible to admins, sales managers, and sales executives.",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "title",
                                    "due_date"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "lead_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Send follow-up proposal",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Email the revised pricing.",
                                        "nullable": true
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "medium",
                                        "nullable": true,
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending",
                                        "nullable": true,
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed"
                                        ]
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01"
                                    },
                                    "reminder_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden for the current role.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Forbidden."
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "/api/leads?page=1",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "example": "/api/leads?page=5",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "/api/leads?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "path": {
                        "type": "string",
                        "example": "/api/leads"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 68
                    }
                },
                "type": "object"
            },
...
            "Task": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Send follow-up proposal"
                    },
                    "description": {
                        "type": "string",
                        "example": "Email the revised pricing.",
                        "nullable": true
                    },
                    "priority": {
                        "type": "string",
                        "example": "medium",
                        "nullable": true,
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "nullable": true,
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed"
                        ]
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-07-01"
                    },
                    "reminder_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserSummary"
                            }
                        ],
                        "nullable": true
                    },
                    "lead": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "first_name": {
                                "type": "string",
                                "example": "John"
                            },
                            "last_name": {
                                "type": "string",
                                "example": "Smith"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "customer": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "company": {
                                "type": "string",
                                "example": "Acme Corp",
                                "nullable": true
                            },
                            "contact_name": {
                                "type": "string",
                                "example": "John Smith"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserSummary": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane.doe@example.com"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the JWT access token obtained from the login or register endpoint, e.g. \"Bearer {token}\".",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}